changed theme and some other stuff

+++ lib/infrastructure/service/sp_service.dart
@@ -2,16 +2,16 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 class SpService {
 
-  Future<void> saveTarget(double fat, double carbs, double protein) async {
+  Future<void> saveTarget(int fat, int carbs, int protein) async {
     final sp = await SharedPreferences.getInstance();
-    sp.setDouble('fat', fat);
-    sp.setDouble('carbs', carbs);
-    sp.setDouble('protein', protein);
+    sp.setInt('fat', fat);
+    sp.setInt('carbs', carbs);
+    sp.setInt('protein', protein);
   }
 
-  Future<({double fat, double carbs, double protein})> loadTarget() async {
+  Future<({int fat, int carbs, int protein})> loadTarget() async {
     final sp = await SharedPreferences.getInstance();
-    final (fat, carbs, protein) = (sp.getDouble('fat') ?? 0, sp.getDouble('carbs') ?? 0, sp.getDouble('protein') ?? 0);
+    final (fat, carbs, protein) = (sp.getInt('fat') ?? 0, sp.getInt('carbs') ?? 0, sp.getInt('protein') ?? 0);
     return (fat: fat, carbs: carbs, protein: protein);
   }
 }
